Create reference-design.md
Pulling copy of the reference design from OpenAPS.org (the version currently live and used for this was last updated Oct. 8, 2015, and was updated to reflect bolus snooze concept). This can be the base for additional updates to it based on changes to the overall design as we move forward; when needed, I will take approved PR versions and paste back to OpenAPS.org to keep it current there.