Update validate-output.md
This pulls in the remaining text suggestions from @oskarpearson's old #45 preparing to loop PR that I closed. This overall could use some more attention in the future by some of our newer loopers (@dm61, @johnmales, @tzarro, @ktomy, others), even if it's just pointing out some of the Qs and we can fill in the answers. Or stub in as much of the answers as you took away from chat sessions, and we can review in the PR process.