Create New-loop-stages-outline.md
High level first draft outline for next generation of docs and approach to setup. Welcoming feedback. Note - Each Stage will have a high level page, each bullet would also be a page underneath. The subbullets underneath a page are just notes about the type of content that'll be included in the page. More info on this approach to come in another few days as I flesh this out in the actual docs to reflect this approach.