Update loops-in-progress.md
It was so hard for me the first time through to submit a PR. So this change is going to be stored on a new branch of my fork Aaybob/docs. I understand that and I think it will show me if I have other changes already in that fork.