Update visualization.md
Added language to clarify that Nightscout integration cannot be completed before creating an openaps instance, setting up devices, reports, etc. Also added a statement that more significant changes to this section are in progress, and pointed to chat. Next: need to go through Ben West's updates, and update the rest of the section accordingly.