Update hardware.md
I changed it from saying which model was the first to not be usable to saying which model was the last to usable. This caused me to buy a pump with firmware 2.5A which I didn't realize until too late that it was just barely too new. I didn't put a letter on the firmware version because I don't know the latest version of 2.4 that was created.