## The determine-basal process suggested change (#257)
When you are told to add the monitor glucose report. the command given just assumed the cgm device was called cgm. I added my device as g5. In the initial adding device section it says name it as you want. So i suggest changing it to the same format as the pump