Add next and previous nav links to page end.
Extend the default Sphinx template to include links to the next and previous documents.  There are small next and previous links at the bottom of the sidebar, but these are easy to miss, and many of the pages are so long that these scroll out of view by the time you reach the end of a page.