Modify Sphinx template to add end-of-page links to previous and next docs in HTML (#336)
* Add next and previous nav links to page end.

Extend the default Sphinx template to include links to the next and previous documents.  There are small next and previous links at the bottom of the sidebar, but these are easy to miss, and many of the pages are so long that these scroll out of view by the time you reach the end of a page.

* Update comments in template

A few more comments to be clearer about what the template does.

* Clean up command syntax for consistency

Changed some of the formatting for consistency with other docs - in particular, remove use of '$' to indicate command prompts, and format with Markdown backticks.  Also, change formatting inside the RST 'note::' blocks - different syntax is necessary there in order to get the same formatting result.

* More Markdown / formatting changes

Some more cleaning up of formatting of command syntax (e.g. hyperlinks, commands in backticks).  Definitely didn't get them all in this document, but it's better.

* Revert "More Markdown / formatting changes"

This reverts commit 6da4ec9e688785bd682bd24e2fcb2c5e4841b0d3.

* Revert "Clean up command syntax for consistency"

This reverts commit 440b36b33afe4c2f81df6830d8182d7f5a3efb9d.