changed description  how to run the setup script
The previous command that was described to run the script didn't work for me (I kept receiving "No such file or directory").  By going to the directory where the script is located first and then type the command to run te script I was however able to run the scripts.