Added step on issue with Jubilinux install (#506)
Added a letter "f" troubleshooting step, moved the final step to "g". The new step includes a link to the depcrecated Ubilinux install instructions in the mmeowlink wiki along with youtube video links. Also, attempted to correct the troubleshooting link just after the Edison flash procedure. Noticed that it said it was linking down the page, when it linked to the mmeowlink wiki.