Description of different Edison variants. (#527)
* Description of different Edison variants.

Added a brief description of different Edison variants/model numbers with some reference material and links to 2 online retailers that allow the buyer to choose which model to purchase. This may be useful as some users have reported heat issues with their Edison modules.

* Added to notes about Edison variants.

Added additional comments to make it clear that the different Edison models have not been tested to determine which perform better or worse for OpenAPS in regard to power use, temperature, or wireless capability.