Update Running Autotune in “manual” mode (#649)
Updating the order of steps. The copies of the profile.json file need to be there before running autotune. Otherwise you get an error. These files are not created automatically when running the setup script if you do not select autotune as one of the advanced features to install. Once the file is copied, autotune runs without any problems