Update on-the-go-wifi-adding.md
Changed the name of the iPhone app to "Termius" instead of "Terminus". When I was going through the instructions I first searched for Terminus but couldn't find it. When I searched for SSH Client I found the Termius app and I assume that's what this doc is referring to. Thanks!