Clarifying some instructions
adding a couple easy troubleshooting, such as for "can't find a PTY" in screen mode.  Removed the ping google.com command because that step right before is using ssh to login.  Would be redundant to run ping if successful at getting in through ssh.  If we want to keep the ping google.com, then we would need to add a line to state control-c to exit the log