ITTT needs to be mentioned up front.
I didn't know how to control the system until I read the doc for days. ITTT should be mentioned at the start for context on reading the rest of the docs. If you accept this change, then the phases may need to be renamed in other parts of the docs.