Update edison.md (#954)
* Update edison.md

The original text said to use M2 as spacers, which means you have to screws these spacers between the boards onto the M2 screw, which is kind of  difficult and impossible to tighten.  If you use slightly larger nuts as spacers, you can just push the screws through the board then through the spacers then the other board and then put a nut on. This is way more secure  and you won't have to tighten them every day!

* Clarify