Patch 3 (#958)
* Updating the setup for Raspberry Pi

This is not close to finished. A lot more for me to do. I will continue in a short while and add a lot of images also.

* Update of Pi instructions

still more to do and pictures to add

* update

still a lot to do

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update Pi-setup.md

* Update to setting up the Raspberry Pi including the Pi Zero W

Please add a quick link to this in the new layout. Many people are still using Pi and quite a few have messaged me for instructions on how to set up Pi Zero W's they have decided they want to move to from now on.

* Formatting