Adding instructions for HTTP Widget App
The link https://github.com/ceben80/http-widget actually DID work, so I've added the instructions here. I didn't erase the other instructions for making an offline webpage, because that method works as well. With both methods, I am having issues with the HTTP widget not refreshing regularly. I also could not figure out how to put in a screenshot. I'll continue to work on those things and add to the docs later.