Steps have changed.
An image needs to be changed and I have the image however I can't upload it to the images folder.  Deleted the connect clicks which are between a couple steps as they seem to be removed from the process.  Will propose this and see what to do with the new image that should replace ![IFTTT maker](../Images/IFTTT_maker.png)