The wifi link gets an error.  Trying to fix...
Needs somebody that knows how to get this to not say you are leaving... ie.. I have the link specified incorrectly or the link is wrong.  However it is a link on wifi debugging so not sure.  Let me know how to fix it correctly and I will.  hmm... there are a few pages that reference the broken link version of the wifi and troubleshooting page.  Need to understand how to reference the correct one and how to be sure I have the correct one (which I tried to link in this doc.)