Update autosens.md (#1085)
* Update all-computers-flash.md (#1056)

Link to the windows standalone driver changed.

* Update update-your-rig.md (#1035)

* Update update-your-rig.md

Is this helpful for some people? Should we update this page to include "Afterward, you also need to turn on the relevant settings in preferences.json." as a reminder for those who have setup and wish to continue using  SMB and UAM?

* Update update-your-rig.md

* Change URL to package install script (#1043)

* re-apply 1/4

* re-apply 2/2 (will likely fail)

* fix attempt 1

* Checking formatting capability

* Fixing formatting back to successful format

* Attempt to > format chart

* Create example-max-safety-chart.md

* Add files via upload

* Add image of chart that won't render in RTD

* Update autosens.md

The paragraph is no longer needed since autosens now recognizes UAM periods.