Update ifttt-integration.md
Yesterday when looking to do Buttons for my bolused carb counts, I could not find the information, I had to ask in Looped. When searching the docs, I did not make logical connection to Low Treatment. I get it now, but really I'm not looking to treat a low, I want to add the carbs I boluses for. Even if same code, I think this will make it easier to find.