Reduce confusion about compatible pumps
I have never before been a Medtronic pump user (always Roche, perhaps that's a European thing). 

When trying to find a pump suitable for use with OpenAPS I was confused about the "How to check pump firmware (check for absence of PC Connect)" section - I had asked the seller of the pump about the "Connect Devices" menu under "Utilities". We both ended up confused.

The intention here is to clarify that the x12, x15 and x22 pumps do not need to have their firmware checked.