Make the use of NS sound a little less optional (#1147)
* Make the use of NS sound a little less optional

(e)SMBs are important, and NS's autotune is required for them. These changes make the importance of NS a little more clear.

* Further clarification

UAM currently requires SMB in 0.6.0, but won't in 0.6.1 (per https://github.com/openaps/oref0/pull/831), so removing that reference and replacing it with a reference to autotune itself.