Update Hardware.md
As promised hereby a new infographic and a short description about read-only pumps. Hopefully you consider it an improvement. If there is anything als it can do please ask. The png is located at my own server. Don't know how to get it at the right place. My english is not flawless. There could be some typos in the text.