Update OpenAPS-install.md
As far as I know, the code just checks if the word "carelink" is part of the punp.ini. I think carelin = on sends the wron signal to the users, since they might guess carelink = off would deactivate it. I suggest to use "radio_type = carelink".
Since https://github.com/openaps/oref0/pull/1011 updated the setup script to do this automatically (but isn't in master yet) I changed the wording to a suggestion to check if it is setup correctly.