Average delta explanation (#1457)
* Average delta explanation

After confusion over what the short_avg_delta and long_avg_delta fields represent to me, an AAPS user.  After much confusion over what was meant by the terms someone pointed me to this document which disagrees with what I was told by a developer.

* Further clarify short_avgdelta and long_avgdelta

The suggested revision was still unclear on how the averaging is performed.  I think this should be unambiguous.