Update troubleshooting for "Monitor/mmtune.json is empty or does not exist" error
I had red and green LEDs on both of my rigs and the "Monitor/mmtune.json is empty or does not exist" error in my loop log. Both rigs were  running dev for several weeks before any problems. I used the above method to fix them both (loops started working again on master branch, have not switched to dev yet). This procedure has obviously not been vetted thoroughly, so it may be of limited use for the docs, but I figured I would document it and it might be a starting point if this is a commonly occurring problem. 

This is my first PR so I tried to integrate it into the existing troubleshooting section to the best of my ability. I also linked to other sections of the docs. I'm not sure how close I came to formatting this correctly, but it's better than inaction I suppose!