use oref0-append-local-temptarget
oref0-append-local-temptarget is the script you want to actually set a local temp target.  If you run oref0-set-local-temptarget, it simply prints out the local target to the screen. oref0-append-local-temptarget is a wrapper script that calls oref0-set-local-temptarget and actually writes the output to a file that will be used as a local temptarget.  My apologies for the confusing names.