Update Codecademy link and add new Python resource
Codecademy's Python page is no longer available at the link that was listed. Updating the reference to a working link.

Also, we are launching Classpert.com, a meta search engine for online courses. We have a collection of Python courses, crawled from providers like Udemy, Coursera, Treehouse, etc.,, so I figured it might be useful for newcomers.