edited wordy sentence
Change the sentence: 'If you get an error, it is likely that you forgot to enter Edison at the first prompt for changing the password.'
Into: 'If you get an error, you likely forgot to enter Edison at the first prompt for changing the password.'
The original sentence is a bit too wordy. The edited sentence is easier to read and clearer.