Put old DIA=3 bilinear information behind a details tag
This isn't anywhere near a clean edit, but a lot of this information is years out of date and contradictory with other parts of the docs. Recreating all of the graphs with accurate to recent version DIAs is a lot of work, but constant disclaimers that graphs do not represent current information would bloat this section a lot. Considering that the vast majority of readers of this page would be on an exponential curve with a DIA of 5 or more, it seems confusing to have that outdated info front and center. Folding to the exponential curve info with a details tag seems better.