Second try at correcting the table
For some reason triple backticks were added around the table, causing it to become a code block instead of a table. I also realized that there were some other inconsistencies in the use of spacing characters, and extra columns on one of the rows. Hopefully this attempt will work better. :crossed_fingers: